[De/Serialization] Scene view window deserialize fix
Working fix for deserializing editor window and other class data that is overwritten after serialization by calling class constructor. Serialize after construction unless a member is required at construction time, in which case serialize before as well as after constructor call. Potentially error prone for construction that alters values that are serialized, but that might be a logic error as such overwritten values (maybe) shouldn't be tagged for serialization.