[Editor] Editor is a console app, and depth rendering fixed (FBO)
Refactoring Editor project to be a console app instead of a static library. Fixed render ordering issue with scene frame buffer rendering by adding a depth texture for FBO to fill. Need to remove dependency on #ifdef _QEDITOR as it won't work for all projects. The solution is moving QE code out of QF domain