[Mirror] Experimenting with type hashing
Less coding today so submitting WIP code. Hoping to deprecate MirrorTypes enum for a has value instead. This removes the need for the enum, and allows "mirroring" or reflecting types "automatically" or in a generic was using templated functions. Removing MIRROR_CLASS_MEMBER_TYPE_OVERRIDE as it was a hack or work around and is now unused.