[ImguiConsole] Autocomplete in previous text field nearly there
Looking to add auto complete hinting. Working in above text output log area but not in input text bar. Imgui seems to hide hints if text bar is non-empty which is not what I want. Tried creating a new function but ran into bgfx imgui library naming conflicts so that has to go as well. Console script comments are in main, and autocomplete non-hints almost done in this commit.