[InputStatesBitRingBuffer] Iterating further on input system specific collection
So initially I just wanted to avoid index bounds checking/correcting by relying on rollover on a unsigned int that had a size of X bits. Then I wanted add pointers to store 2 windows to use for last and current app frames for detecting input this frame and comparing it to last frame. Then I wanted to move the pointer logic to a new class to clean up the input system code. Then I wanted to pair the QKey codes with the matching bool state inside of the same collection, to further encapsulate the logic. Then I wanted to reduce the size of key states from an 8 byte bool to an individual 1 bit value.

The (soon to be) end result is a collection that uses indices of size X bits as provided by user through template, that stores typename T value with matching 0/1 bit states, that provides the option of adding index "markers" for defining ranges within the collection while utilizing a head pointer for writing. So 1 direction a ring buffer, without bounds checks, using minimal RAM and encapsulating as much logic internally as possible. Maybe a waste of time, but a fun concept to explore so far.