[EditorOnKeyCallback] Moving Editor key event logic into a new function that is called on key callback
Looking to reduce the amount of polling being done for editor keys. As new input mappings are introduced, many new hotkeys will be added and the amount of input checks may skyrocket so I'd like to get ahead of the curve and swapping unnecessary polling to at least only poll on key event.